third_party/blink/web_tests/media/video-defaultmuted.html

<!DOCTYPE html>
<title>Test "defaultMuted" IDL, "muted" IDL and content attributes.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
    var video = document.createElement("video");
    // Test that "defaultMuted" IDL attribute reflects "muted" content attribute and
    // that "muted" IDL attribute is not affected by content attribute.
    assert_false(video.muted);
    assert_false(video.defaultMuted);
    video.setAttribute("muted", "muted");
    assert_true(video.muted);
    assert_true(video.defaultMuted);
    video.removeAttribute("muted", "muted");
    assert_false(video.muted);
    assert_false(video.defaultMuted);

    // Test that "muted" IDL attribute is not affected by "defaultMuted" IDL attribute.
    video.defaultMuted = true;
    assert_true(video.muted);
}, "Test 'defaultMuted' and 'muted' IDL attributes");
</script>
